refactor(authz): split module, make global by default
extract core module from authz, rename options,
remove `isGlobal` option
