+++ packages/nestjs-authorization/src/interfaces/authorization-module.interface.ts
@@ -2,21 +2,22 @@ import { Abstract, ModuleMetadata, Provider, Type } from '@nestjs/common';
 import { PermissionProvider } from '../permission.provider';
 
 export interface AuthorizationOptionsFactory {
-  createOptions(): Promise<AuthorizationOptions> | AuthorizationOptions;
+  createOptions():
+    | Promise<AuthorizationModuleOptions>
+    | AuthorizationModuleOptions;
 }
 
-export interface AuthorizationOptions {
+export interface AuthorizationModuleOptions {
   permissionProvider: PermissionProvider;
 }
 
 export interface AuthorizationModuleAsyncOptions
   extends Pick<ModuleMetadata, 'imports'> {
-  isGlobal?: boolean;
   inject?: Array<Type<any> | string | symbol | Abstract<any> | Function>;
   extraProviders?: Provider[];
   useExisting?: Type<AuthorizationOptionsFactory>;
   useClass?: Type<AuthorizationOptionsFactory>;
   useFactory?: (
     ...args: any
-  ) => AuthorizationOptions | Promise<AuthorizationOptions>;
+  ) => AuthorizationModuleOptions | Promise<AuthorizationModuleOptions>;
 }
